requirements.txt
customtkinter
openmc
numpy
tk
matplotlib
Pillow